Backend/requirements.txt
aiofiles
aiohappyeyeballs
aiohttp
aiohttp-retry
aiosignal
aiosmtplib
annotated-types
anyio
argon2-cffi
argon2-cffi-bindings
asgiref
asttokens
attrs
audioop-lts
backoff
bcrypt
beautifulsoup4
blinker
build
cachetools
certifi
cffi
charset-normalizer
chroma-hnswlib
chromadb
click
colorama
coloredlogs
cryptography
dataclasses-json
decorator
Deprecated
diffusers
distro
dnspython
durationpy
ecdsa
email_validator
executing
fastapi
fastapi-mail
fastapi-users
fastapi-users-db-sqlalchemy
fastjsonschema
filelock
filetype
flatbuffers
frozenlist
fsspec
generativeai
google
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-generativeai
google-oauth
googleapis-common-protos
greenlet
grpcio
grpcio-status
h11
httpcore
httplib2
httptools
httpx
httpx-sse
huggingface-hub
humanfriendly
idna
importlib_metadata
importlib_resources
ipython
itsdangerous
jedi
Jinja2
jiter
joblib
jose
jsonpatch
jsonpointer
kubernetes
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-openai
langchain-text-splitters
langsmith
makefun
markdown-it-py
MarkupSafe
marshmallow
matplotlib-inline
mdurl
mmh3
monotonic
mpmath
multidict
mypy-extensions
networkx
numpy
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-cusparselt-cu12
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvtx-cu12
oauthlib
onnxruntime
openai
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation
opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-fastapi
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
opentelemetry-util-http
orjson
overrides
packaging
parso
passlib
pexpect
pillow
posthog
prompt_toolkit
propcache
proto-plus
protobuf
ptyprocess
pure_eval
pwdlib
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic-settings
pydantic_core
Pygments
PyJWT
PyMySQL
pyOpenSSL
pyparsing
PyPika
pyproject_hooks
python-dateutil
python-dotenv
python-jose
python-multipart
PyYAML
redis
regex
requests
requests-oauthlib
requests-toolbelt
rich
rsa
safetensors
scikit-learn
scipy
sentence-transformers
setuptools
shellingham
six
sniffio
soupsieve
SpeechRecognition
SQLAlchemy
stack-data
standard-aifc
standard-chunk
starlette
sympy
tabulate
tenacity
threadpoolctl
tiktoken
tokenizers
torch
tqdm
traitlets
transformers
triton
twilio
typer
typing-inspect
typing_extensions
uritemplate
urllib3
uvicorn
uvloop
watchfiles
wcwidth
websocket-client
websockets
wrapt
yarl
zipp
zstandard
alembic
google-api-python-client
accelerate